requirements.txt
cryptography==3.4.8
dnspython==2.1.0
matplotlib==3.3.4
numpy==1.20.1
argparse==1.1
re==2.2.1
